// src/app/shared/models/gbm_blok.model.ts
export class GbmBlok {
	id?: number;
  kode?: string;
  nama?: string;
  organisasi_id?:number;
  tahuntanam?:number;
  luasareaproduktif?:number;
  luasareanonproduktif?:number;
  jumlahpokok?:number;
  statusblok?: string;
  topografi?: string;
  klasifikasitanah?: string;
  jenisbibit?: string;
  intiplasma?: string;
  kelaspohon?: any;
  tahunmulaipanen?: any;
  bulanmulaipanen?: any;
  kodetanah?: any;
  tanggaltransaksi?: any;
  tanggalpengakuan?: any;
  basiskg?: any;
  periodetm?: any;
  cadangan?: any;
  okupasi?: any;
  rendahan?: any;
  sungai?: any;
  rumah?: any;
  kantor?: any;
  pabrik?: any;
  jalan?: any;
  kolam?: any;
  umum?: any;
  lc?: any;

}